Display roll queue elements conditionally
Based on how many skills components are added to the roll queue, the User should see different UI elements (like a display of the multi-action penalty, or only displaying the "remove skills" button if there are any skills in the queue).  Also added logic for rolling botches and handling the multi-action penalty for roll results.