Utilize grid to better handle height of elements
In short, I utilized two grids to get the skill list & roll queue to match the height of the browser; first, a column grid that contained the character details and a second grid container as the second child.  This second grid contained the skills and roll queue.  I applied flex sizing to everything, and it's all dynamic now.  I didn't do anything for improving the resizing of individual skills or the roll queue skills, nor did I implement any handling for mobile.  Felt more important to move onto routing!