Add character selection for skill roller
User is able to select the character they'd like to roll skills for on the skill-roller tab.  The skills are also populated from the character object now instead of being pulled from a static skills object in the DB.  I am running into issues with refreshing or navigating directly to a character (i.e. /skill-roller/axel-prose); this is because that component makes a call to get character info, and really that should live in a parent component.  Rather than refactor that stuff prematurely, I'm going to implement Redux for state management and do it once.