Implement async Redux action for pulling characters
Although messy, there is now a Redux store in place, and character data is being fetched asynchronously and placed in the store.  When the User selects a character, they are saved to the state as the selected character; however, I'm not checking for a selected character on the skill roller tab, so navigating away and returning still forces the User to select a character again.  Given the rush to implement something, I'm pushing this and moving on.